modified doc strings and function names for easy usage  nexB#2
