--- src/README.rst
@@ -1,2 +1,0 @@
-Put your Python source code (and installable data) in this directory.
-
