+++ src/scorecode/__init__.py
@@ -1,8 +1,11 @@
 #
 # Copyright (c) nexB Inc. and others. All rights reserved.
-# ScanCode is a trademark of nexB Inc.
 # SPDX-License-Identifier: Apache-2.0
 # See http://www.apache.org/licenses/LICENSE-2.0 for the license text.
-# See https://github.com/nexB/scancode-toolkit for support or download.
+# See https://github.com/aboutcode-org/scorecode for support or download.
 # See https://aboutcode.org for more information about nexB OSS projects.
 #
+
+OSSF_SCORECARD_API_URL = "https://api.securityscorecards.dev"
+
+
